# swagger.yaml
openapi: 3.0.0
info:
  title: Car Contract Management System API
  description: 중고차 계약 관리 시스템 API 문서
  version: 1.0.0
  contact:
    name: API Support
    email: support@carmate.com
servers:
  - url: http://localhost:${process.env.PORT || 3000}
    description: Development server
    # 배포 전이므로 주석 처리
  # - url: ${process.env.PROD_SERVER_URL || 'https://api.carmate.com'}
  #  description: Production server
...
security:
  - bearerAuth: []
paths:
...
  /contractDocuments/{contractDocumentId}/download:
    get:
      tags:
        - ContractDocuments
      summary: 계약서 다운로드
      description: 단일 계약서 파일 다운로드
      parameters:
        - in: path
          name: contractDocumentId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 계약서 다운로드 성공
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /contractDocuments/{contractId}:
    patch:
      tags:
        - ContractDocuments
      summary: 계약서 수정
      description: 계약서 추가/삭제
      parameters:
        - in: path
          name: contractId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                deleteDocumentIds:
                  type: array
                  items:
                    type: integer
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '200':
          description: 계약서 수정 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: '계약서 수정 성공'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /contractDocuments:
    get:
      tags:
        - ContractDocuments
      summary: 계약서 목록 조회
      description: 등록된 계약서 목록 조회
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - in: query
          name: searchBy
          schema:
            type: string
            enum: [contractName, userName]
        - $ref: '#/components/parameters/KeywordParam'
      responses:
        '200':
          description: 계약서 목록 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractDocumentListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /contractDocuments/upload:
    post:
      tags:
        - ContractDocuments
      summary: 계약서 업로드
      description: 계약서 파일 업로드 (최대 10개, 각 파일 최대 10MB)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - contractId
                - files
              properties:
                contractId:
                  type: integer
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                  maxItems: 10
      responses:
        '200':
          description: 계약서 업로드 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: '계약서 업로드 성공'
                  contractDocumentId:
                    type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /contractDocuments/download:
    post:
      tags:
        - ContractDocuments
      summary: 계약서 다중 다운로드
      description: 여러 계약서를 ZIP 파일로 다운로드
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - contractDocumentIds
              properties:
                contractDocumentIds:
                  type: array
                  items:
                    type: integer
      responses:
        '200':
          description: 계약서 다운로드 성공
          content:
            application/zip:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      in: query
      name: page
      schema:
        type: integer
        default: 1
      description: 페이지 번호
    PageSizeParam:
      in: query
      name: pageSize
      schema:
        type: integer
        default: 8
      description: 페이지당 아이템 수
    KeywordParam:
      in: query
      name: keyword
      schema:
        type: string
      description: 검색 키워드
  schemas:
...
    ContractDocumentListResponse:
      type: object
      properties:
        currentPage:
          type: integer
        totalPages:
          type: integer
        totalItemCount:
          type: integer
        data:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              contractName:
                type: string
              resolutionDate:
                type: string
                format: date
              documentsCount:
                type: integer
              manager:
                type: string
              carNumber:
                type: string
              documents:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    fileName:
                      type: string

    # Dashboard schemas
...
    Error:
      type: object
      properties:
        message:
          type: string
        code:
          type: string
        timestamp:
          type: string
          format: date-time
  responses:
    BadRequest:
      description: 잘못된 요청
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: '잘못된 요청입니다'
            code: 'BAD_REQUEST'
            timestamp: '2024-01-01T00:00:00Z'
    Unauthorized:
      description: 인증이 필요합니다
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: '로그인이 필요합니다'
            code: 'UNAUTHORIZED'
            timestamp: '2024-01-01T00:00:00Z'
...
    NotFound:
      description: 리소스를 찾을 수 없습니다
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: '존재하지 않는 리소스입니다'
            code: 'NOT_FOUND'
            timestamp: '2024-01-01T00:00:00Z'
...